% Generated by roxygen2 (4.0.0): do not edit by hand
\name{added.var}
\alias{added.var}
\title{Added variable plots of predictors}
\usage{
added.var(Y, X, main = "")
}
\arguments{
\item{Y}{a vector representing a single response}

\item{X}{two or more predictor columns in a matrix or dataframe}

\item{main}{subtitle for the plot}
}
\value{
a plot is sent to the current graphics device (no value is returned)
}
\description{
Produces an added variable plot given 1 response and 2 or more predictors.
}
\examples{
added.var(Y=mtcars$hp, X=mtcars[, c("mpg", "disp", "wt")], main="Predicting horsepower from MPG, displacement, and weight")
}

