% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addedvar}
\alias{addedvar}
\title{Added Variable Plots of Predictors}
\usage{
addedvar(Y, X, main = "")
}
\arguments{
\item{Y}{A vector representing a single response.}

\item{X}{Two or more predictor columns in a matrix or dataframe.}

\item{main}{Subtitle for the plot.}
}
\value{
A plot is sent to the current graphics device (no value is returned).
}
\description{
Produces an added variable plot given 1 response and 2 or more predictors.
}
\examples{
addedvar(Y=mtcars$hp, X=mtcars[, c("mpg", "disp", "wt")], main="Predicting horsepower from MPG, displacement, and weight")
}

