% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addhist}
\alias{addhist}
\title{Add a Histogram}
\usage{
addhist(x, y = NULL, type = "xy", nclass = 20, newmar = 0:1,
  adj.fac = 1.05, xlab = "Frequency", ylab = "", fill = "gray")
}
\arguments{
\item{x}{A numeric vector, the first set of data to be binned into a histogram.}

\item{y}{A numeric vector, the second set of data to be binned into a histogram, default NULL.}

\item{type}{A character scalar, indicating to which plot axes histograms should be added; \code{"x"} adds a histogram along the x-axis,
\code{"y"} adds a histogram along the y-axis, \code{"xy"} (the default) adds a histogram along both axes.}

\item{nclass}{An integer scalar, the target number of bins for the histogram, default 20.}

\item{newmar}{A numeric vector of length 2, indicating new margins to use, default c(0, 1).}

\item{adj.fac}{A numeric scalar, adjustment factor for extent of y-axis, default 1.05 (to ensure tallest bars in histogram are below axis box).}

\item{xlab}{A character scalar, label for x-axis, default "Frequency".}

\item{ylab}{A character scalar, label for y-axis, default "".}

\item{fill}{A character or numeric scalar, color for filling in histogram bars, default "gray".}
}
\description{
Add a marginal histogram to a plot.
}
\examples{
# fake data
xx <- rnorm(30)
yy <- runif(30)

# type "x"
layout(matrix(2:1, ncol=1), heights=c(1/5, 4/5))
par(mar=c(4, 4, 0, 1), cex=1, las=1)
plot(xx, yy)
addhist(xx, type="x")

# type "y"
layout(matrix(1:2, ncol=2), widths=c(4/5, 1/5))
par(mar=c(4, 4, 1, 0), cex=1, las=1)
plot(xx, yy)
addhist(yy, type="y")

# type "xy"
layout(matrix(c(2, 1, 0, 3), ncol=2), heights=c(1/5, 4/5), widths=c(4/5, 1/5))
par(mar=c(4, 4, 0, 0), cex=1, las=1)
plot(xx, yy)
usr <- par("usr")
mar <- par("mar")
addhist(xx, yy, type="xy")
}
\seealso{
\code{\link{hist}}.
}

