% Generated by roxygen2 (4.0.0): do not edit by hand
\name{AICc}
\alias{AICc}
\title{Comparison of models using Akaike Information Criterion}
\usage{
AICc(fitlist, corr = TRUE)
}
\arguments{
\item{fitlist}{a list of model fits to compare, e.g., lm/glm/aov objects}

\item{corr}{a logical indicating whether the AIC should be corrected for small sample size, defaults to TRUE}
}
\value{
data frame with a row for each model being compared,
ordered by either the uncorrected AIC (corr=FALSE) or the AIC corrected for small sample size (corr=TRUE).
Columns include the number of observations (n), the number of parameters (p), the root mean squared error (rmse),
the uncorrected AIC (aic), the AIC corrected for small sample size (aicc),
the delta AIC (daic or diacc), and the weights of evidence (aicw or aiccw).
}
\description{
Compares a collection of statistical models using AIC.
}
\examples{
fit1 <- lm(hp ~ mpg + disp + wt, data=mtcars)
fit2 <- lm(hp ~ mpg + disp, data=mtcars)
fit3 <- lm(hp ~ mpg + wt, data=mtcars)
fit4 <- lm(hp ~ disp + wt, data=mtcars)
AICc(list(fit1, fit2, fit3, fit4))
}

