% Generated by roxygen2 (4.0.0): do not edit by hand
\name{all.combsb}
\alias{all.combsb}
\title{All combinations bbb}
\usage{
\method{all}{combsb}(num, from = 0, to = num)
}
\arguments{
\item{num}{a scalar, the number of items}

\item{from}{a scalar, the minimum number of items in each combination, default 0}

\item{to}{a scalar, the maximum number of items in each combination, default \code{num}}
}
\value{
a matrix with rows corresponding to each possible combination and columns corresponding to item number
}
\description{
All possible combinations of the given number of items.
}
\examples{
all.combs(3)
}

