% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignReviewer.R
\name{assignReviewer}
\alias{assignReviewer}
\title{Assign Reviewers to Papers}
\usage{
assignReviewer(data, Unitvar, prefOrd = c("LSBS", "HBBS", "LMERS", "Coastal",
  "LMichigan", "LHuron", "LEBS", "LOBS", "TLAS"))
}
\arguments{
\item{data}{Data frame of papers to be reviewed, each row represents a single paper.}

\item{Unitvar}{Character scalar, name of variable in \code{data} identifying the Unit with
which each paper is associated.}

\item{prefOrd}{Character vector, preferred ordering of Units, default geographical,
\code{c("LSBS", "HBBS", "LMERS", "Coastal", "LMichigan", "LHuron",
"LEBS", "LOBS", "TLAS")}.}
}
\value{
A data frame, the same as \code{data} with one new column, \code{Reviewer},
  and sorted by Reviewer.
}
\description{
Given a data frame of papers, in which each paper is associated with a
Unit, assign a reviewing Unit to each paper such that each reviewing unit
has approximately the same number of papers to review and no reviewing unit
is assigned one of its own papers to review.
}
\details{
If a vector of length 1 is provided for \code{round2}, \code{ndec},
  \code{comma}, or \code{align}, the same rounding, decimals, commas, or
  alignment is applied to all specified columns.
}
\examples{
test <- data.frame(Source=rep(paste("Group", 1:5), c(2, 3, 1, 6, 4)),
 Paper=LETTERS[1:16])
test2 <- assignReviewer(data=test, Unitvar="Source",
 prefOrd=paste("Group", 1:5))
test
test2
with(test2, addmargins(table(Reviewer, Source)))
}
\seealso{
\code{\link{format}}.
}
