% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bcpCI.r
\name{bcpCI}
\alias{bcpCI}
\title{Bias-Corrected Percentile Confidence Interval}
\usage{
bcpCI(tboot, orig, alpha = 0.05)
}
\arguments{
\item{tboot}{A numeric vector of bootstrap estimates, typically \code{$t} from the output of the \code{\link{boot}} function.}

\item{orig}{A numeric scalar, the original estimate from the data, typically \code{$t0} from the output of the \code{\link{boot}} function.}

\item{alpha}{A numeric scalar, the desired significance level for 100*(1-\code{alpha})\% confidence limits, default 0.05.}
}
\value{
A named numeric vector of length 2, with the lower and upper confidence limits.
}
\description{
Calculates the bias-corrected percentile confidence interval from a bootstrap sample.
}
\examples{
bcpCI(exp(rnorm(20)), 1)

\dontrun{
# Bootstrap of the ratio of means using the city data
library(boot)
ratio <- function(d, w) sum(d$x * w)/sum(d$u * w)
results <- boot(city, ratio, R=999, stype="w")
results$t0
bcpCI(results$t, results$t0, 0.1)
}
}
\references{
Manly, Bryan F. J.  1997.
	Randomization, Bootstrap and Monte Carlo Methods in Biology.  Chapman & Hall, London.
}

