% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brewcol.R
\name{brewcol}
\alias{brewcol}
\title{Assign Colors to Collection of Values Using Color Brewer}
\usage{
brewcol(x, n = 9, name = c("YlGnBu", "Oranges", "Greens", "Blues", "Greys",
  "Dark2")[1])
}
\arguments{
\item{x}{A numeric vector of values to which colors will be assigned.}

\item{n}{An integer scalar indicating the number of unique colors to be assigned,
default 9.  The range of possible values depends on the palette named,
typically \code{n} should be from 3 to 9.}

\item{name}{A character scalar indicating the palette name to be passed to
\code{\link{brewer.pal}}, default "YlGnBu".}
}
\value{
A character vector of colors expressed as hexadecimal values with a "#"
  prefix.
}
\description{
Assign a specified number of Color Brewer colors to a collection of values.
}
\examples{
x <- 1:20
mycol <- brewcol(x, 4, "Dark2")
plot(x, x, col=mycol, pch=16, cex=3)
}

