% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcr2}
\alias{calcr2}
\title{Coefficient of determination}
\usage{
calcr2(fitted, observed, nparam)
}
\arguments{
\item{fitted}{a vector, fitted values}

\item{observed}{a vector, observed values}

\item{nparam}{a scalar, number of parameters in the fitted model}
}
\value{
vector, unadjusted and adjusted coefficients of determination
}
\description{
Calculate the unadjusted and adjusted coefficient of determination (R^2).
}
\examples{
fit1 <- lm(mpg ~ cyl + disp, data=mtcars)
summary(fit1)
calcr2(fit1$fitted, mtcars$mpg, 3)
fit2 <- lm(mpg ~ cyl + disp + wt, data=mtcars)
summary(fit2)
calcr2(fit2$fitted, mtcars$mpg, 4)
}

