% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capwords.r
\name{capwords}
\alias{capwords}
\title{Capitalize}
\usage{
capwords(s, strict = TRUE)
}
\arguments{
\item{s}{A vector of strings.}

\item{strict}{A logical indicating whether other letters should be converted to lower case, default TRUE.}
}
\value{
A vector the same length as \code{s}.
}
\description{
Capitalize the first letter of every word.
}
\examples{
capwords(c("using AIC for model selection"))
capwords(c("using AIC", "for MODEL selection"), strict=FALSE)
}
\seealso{
\code{\link{casefold}}, from which the function was derived.
}

