% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/char2num.R
\name{char2num}
\alias{char2num}
\title{Convert Character to Numeric}
\usage{
char2num(x, varname = NULL, pmissing = c("NA", ".", "", " ", "-"),
  justindex = FALSE)
}
\arguments{
\item{x}{A character vector to be converted to numeric.}

\item{varname}{A character scalar to be used as an identifier in the generated
comment vector.  If NULL (default), the name of x will be used.
If FALSE, no varname will be used in the comment.}

\item{pmissing}{A character vector of all possible values to be interpreted as missings
(and, therefore, not be commented upon).}

\item{justindex}{A logical scalar indicating if the output should just be the index of
non-numeric values, default FALSE.}
}
\value{
By default, a list of length two with the converted numeric vector and
  the comment vector.
	 If justindex=TRUE, a character vector the same length as x,
  indicating the elements with non-numeric values.
}
\description{
Convert a character vector to a numeric vector, save any non-numeric values
to a separate vector of comments.
}
\examples{
a <- c(">1.5", "3", "NA", "missing", ".", "12")
char2num(a)
char2num(a, FALSE)
char2num(a, "pH")
char2num(a, pmissing=c("NA", "missing", "."))
char2num(a, justindex=TRUE)

df <- data.frame(
	x1=c(NA, 8.2, 8.2, 8.2, 8.6, 8.1, 8.2),
	x2=c(NA, "83", "*", "83", "79-80", "NA", "83"),
 x3=c(NA, NA, NA, "9.4", "?", ">10", "6.6"))
vars <- c("x2", "x3")
tonum <- lapply(vars, function(v) char2num(df[, v], varname=v))
df2 <- df
df2[, vars] <- data.frame(sapply(tonum, "[", 1))
df2$comment <- apply(do.call(cbind, sapply(tonum, "[", 2)), 1,
 paste, collapse="")
}

