% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chi}
\alias{chi}
\title{Chi-Squared Test}
\usage{
chi(x, rpct = 0)
}
\arguments{
\item{x}{A numeric matrix.}

\item{rpct}{A numeric scalar indicating the rounding used for printed output, default 0.}
}
\value{
A list with class "htest" containing the components described in \link{\code{chisq.test}}.
}
\description{
Performs chi-squared contingency table tests with informative output.
}
\examples{
## From Agresti(2007) p.39
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(gender = c("M", "F"), party = c("Democrat", "Independent", "Republican"))
chi(M)
}
\seealso{
\link{\code{chisq.test}}.
}

