% Generated by roxygen2 (4.0.1): do not edit by hand
\name{circles}
\alias{circles}
\title{Draw Circles on a Plot}
\usage{
circles(x, y, z, data.range = range(z, na.rm = TRUE),
  circle.size.range = c(0.1, 1), outx = NA, outy = NA, add = FALSE,
  xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{Numeric vector of x coordinates.}

\item{y}{Numeric vector of y coordinates.}

\item{z}{Numeric vector of data used to generate circles.}

\item{data.range}{Numeric vector, length 2, minimum and maximum z data to plot, default \code{range(z, na.rm=TRUE)}.}

\item{circle.size.range}{Numeric vector, length 2, minimum and maximum circle radii in inches, default 0.1 to 1.}

\item{outx}{Numeric scalar of x coordinate beyond the figure margins, default NA (see details).}

\item{outy}{Numeric scalar of y coordinate beyond the figure margins, default NA (see details).}

\item{add}{Logical scalar specifying if circles are added to existing plot (TRUE), default FALSE (a new plot is created).}

\item{xlim}{Numeric vector, length 2, x-axis limits, unused if \code{add=TRUE}.}

\item{ylim}{Numeric vector, length 2, y-axis limits, unused if \code{add=TRUE}.}

\item{...}{Additional parameters supplied to the \code{\link{symbols}} function.}
}
\value{
A data frame with the name, class, dimension, and size of each member of the environment.
}
\description{
Draw circles on a plot, with control over circle size.
}
\details{
The size of the circles plotted corresponds directly with the range of data.
For example, if there is a z of size \code{data.range[1]}, it will be plotted as a circle with radius \code{circle.size.range[1]},
and if there is a z of size \code{data.range[2]}, it will be plotted as a circle with radius \code{circle.size.range[2]}.

The default of NA for \code{outx} and \code{outy} places unseen smallest and biggest circles at a location
where the x and y coordinates are 10 times the range observed plus the maximum observed.
In most instances this should be well beyond the figure margins.
}
\examples{
circles(trees$Height, trees$Girth, trees$Volume, data.range=sqrt(c(0, max(trees$Volume))),
	circle.size.range=c(0, 0.3), xlab="Height (ft)", ylab="Diameter (in)", main="Tree Volume")
}

