% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{Quick Clean Up}
\usage{
cleanup()
}
\description{
Removes all objects from the current working directory.
}
\details{
User is prompted for response.  If the first letter of the response is "y" or "Y", all objects are removed from the current working directory.
}

