% Generated by roxygen2 (4.0.2): do not edit by hand
\name{colr}
\alias{colr}
\title{Create a Range of Colors}
\usage{
colr(x, fromcolname, tocolname)
}
\arguments{
\item{x}{A numeric vector, the values to be assigned colors.}

\item{fromcolname}{A character or numeric scalar, indicating the color to use for the lowest value in \code{x}.
Either a color name (as listed by \code{\link{colors}()},
a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see \code{\link{rgb}}),
or a positive integer i meaning \code{\link{palette}()[i]}.}

\item{tocolname}{A character or numeric scalar, indicating the color to use for the highest value in \code{x}.
See \code{fromcolname}.}
}
\description{
Create a range of colors between two specified colors.
}
\examples{
x <- 1:10
plot(x, x, pch=16, col=colr(x, "blue", "yellow"), cex=4)
}
\seealso{
\code{\link{colors}}, \code{\link{palette}}.
}

