% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combLevel.R
\name{combLevel}
\alias{combLevel}
\title{Combine Levels}
\usage{
combLevel(n)
}
\arguments{
\item{n}{An integer scalar, the number of levels.}
}
\value{
A matrix with \code{n} columns with one row for each possible combination
  of the \code{n} levels.
}
\description{
Generates all possible combinations of a given number of levels.
}
\examples{
combLevel(3)
combLevel(4)
}
\references{
Based on a method posted to the Adamistics blog 1 March 2016
  \href{http://adamistics.blogspot.com/2016/03/generating-combinations-of-levels.html}{[link]}.
}

