% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coordflip}
\alias{coordflip}
\title{Flip Coordinates}
\usage{
coordflip(pts, seg)
}
\arguments{
\item{pts}{A matrix or data frame with the first two columns of numeric x and y coordinates to be rotated.}

\item{seg}{A matrix or data frame of dimension 2 x 2 giving two points which define the line over which the coordinates will be flipped.}
}
\value{
A matrix or data frame with two columns of flipped x and y coordinates.
}
\description{
Flip (or reflect) coordinates across any given line.
}
\examples{
# starting coordinates
test <- matrix(c(0, 4, 1, 0, 2, 3), ncol=2, dimnames=list(LETTERS[1:3], NULL))
coordplot(test)
# flip the coordinates across the line defined by the first two points
ftest <- coordflip(test, test[1:2, ])
coordplot(ftest)
}
\references{
Based on a method posted by \strong{Il-Bhima} on 22 July 2010 on
\href{http://stackoverflow.com/questions/3306838/algorithm-for-reflecting-a-point-across-a-line}{stackoverflow}.
}
\seealso{
\code{\link[jvamisc]{coordplot}}, \code{\link[jvamisc]{coordmove}}, \code{\link[jvamisc]{coordturn}}.
}

