% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordplot.R
\name{coordplot}
\alias{coordplot}
\title{Plot Coordinates}
\usage{
coordplot(pts)
}
\arguments{
\item{pts}{A numeric matrix with two columns of x and y coordinates to be plotted.}
}
\description{
Quick plot of coordinates on equally scaled coordinate plane,
labelled with row names.
}
\examples{
test <- matrix(c(0, 4, 1, 0, 2, 3), ncol=2,
 dimnames=list(LETTERS[1:3], NULL))
coordplot(test)
}
\seealso{
\code{\link[jvamisc]{coordturn}}, \code{\link[jvamisc]{coordmove}},
  \code{\link[jvamisc]{coordflip}}, \code{\link[jvamisc]{coordtri}}.
}

