% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coordturn.r
\name{coordturn}
\alias{coordturn}
\title{Rotate Coordinates}
\usage{
coordturn(pts, pvt, rot)
}
\arguments{
\item{pts}{A numeric matrix with two columns of x and y coordinates to be rotated.}

\item{pvt}{A numeric vector of length 2, the x and y coordinates of the pivot point around which \code{pts} will be rotated.}

\item{rot}{A numeric scalar indicating the amount of clockwise rotation, in radians.}
}
\value{
A numeric matrix with same dimension as \code{pts} with the rotated x and y coordinates.
}
\description{
Rotate coordinates clockwise around a pivot point.
}
\examples{
# starting coordinates
test <- matrix(c(0, 4, 1, 0, 2, 3), ncol=2,
	dimnames=list(LETTERS[1:3], NULL))
coordplot(test)
# rotate the coordinates clockwise 45 degrees
# around the first point (the origin)
rottest <- coordturn(test, test[1, ], rot=pi/4)
coordplot(rottest)
# rotate the coordinates counterclockwise 45 degrees
# around the first point (the origin)
rottest <- coordturn(test, test[1, ], rot=-pi/4)
coordplot(rottest)
}
\references{
Modification of code posted by Sage on 3 March 2011 on the website benn.org
\href{http://benn.org/2007/01/06/rotating-coordinates-around-a-centre/}{[link]}.
}
\seealso{
\code{\link[jvamisc]{coordplot}}, \code{\link[jvamisc]{coordmove}}, \code{\link[jvamisc]{coordflip}}, \code{\link[jvamisc]{coordtri}}.
}

