% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-data.R
\name{counties}
\alias{counties}
\title{United States County Boundaries Map}
\format{A list (of class map) with elements x, y, range, and names.
  The x and y vectors have longitude (x) and latitude (y) coordinates of
  successive county polygons, separated by NAs.}
\source{
A copy of the \code{county} data base from the \code{maps} package.
}
\description{
Geographical data base for the county boundaries of the US,
  generated from US Department of the Census data.
}

