% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfclip}
\alias{dfclip}
\title{Read Data Frame from Clipboard}
\usage{
dfclip(...)
}
\arguments{
\item{...}{Additional parameters to \link{\code{read.table}}.}
}
\value{
Data frame.
}
\description{
Read in a data frame from text pasted to the clipboard.
}
\seealso{
\link{\code{read.table}}.
}

