% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy.R
\name{doy}
\alias{doy}
\title{Day of Year}
\usage{
doy(date, day1 = "01-01")
}
\arguments{
\item{date}{A vector of dates to be converted.}

\item{day1}{A character scalar specifying what date should equate to day 1,
use "01-01", the default, to get the Julian day.}
}
\value{
A numeric vector the same length as \code{date} giving the day of the year.
}
\description{
Calculates the day of the year.
}
\examples{
x <- as.Date(c("1963-01-15", "1972-07-20", "1999-03-10"))
doy(x)
doy(x, "03-01")
}

