% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawcell.R
\name{drawcell}
\alias{drawcell}
\title{Draw Table Cell}
\usage{
drawcell(title, r, c, text.cex = 1, bg.col = "white", frame.cell = TRUE)
}
\arguments{
\item{title}{A character scalar, the text to be placed in the table cell.}

\item{r}{An integer scalar, the row number of the table cell.}

\item{c}{An integer scalar, the column number of the table cell.}

\item{text.cex}{A numeric scalar, the relative size of the text to be place in
the table cell, default 1.}

\item{bg.col}{A character or numeric scalar, indicating the background color to use
for the table cell, default "white".}

\item{frame.cell}{A logical scalar, indicating whether a black border should be drawn
around the table cell, default TRUE.}
}
\value{
A table cell is added to the current table.
}
\description{
Draw a table cell at location (r,c).
}
\details{
This is part of group of function to plot tables, modifications of
  functions in the Systmatic Investor Toolbox.
}
\references{
Systmatic Investor Toolbox.
  \emph{https://github.com/systematicinvestor/SIT}
}
