% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/endrtf.R
\name{endrtf}
\alias{endrtf}
\title{Write and Close an RTF Document}
\usage{
endrtf(rtf = doc, details = FALSE, ...)
}
\arguments{
\item{rtf}{An rtf object, default \code{doc}.}

\item{details}{Logical scalar indicating if session details should be added to
the end of the document, default FALSE.}

\item{...}{Additional parameters to \code{\link[rtf]{addPageBreak}}.}
}
\description{
Write and close an rtf (rich text format) document.
}
\examples{
\dontrun{
doc <- startrtf()
heading("Heading 1")
para("First paragraph.")
tab <- head(cars)
tabl("First few rows of cars data.", row.names=FALSE)
heading("Heading 2", 2)
para("Second paragraph.")
fig <- function() {
	plot(cars)
	lo <- loess(cars$dist ~ cars$speed)
	lines(lo$x, lo$fitted)
}
figu("Speed vs. distance from the cars data.")
endrtf()
}
}
\seealso{
\code{\link[jvamisc]{startrtf}}, \code{\link[jvamisc]{heading}},
  \code{\link[jvamisc]{para}}, \code{\link[jvamisc]{tabl}},
  \code{\link[jvamisc]{figu}}, \code{\link[jvamisc]{figbig}},
  \code{\link[rtf]{RTF}}, \code{\link[rtf]{addPageBreak}}.
}

