% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fill}
\alias{fill}
\title{Fill in Missing Values}
\usage{
fill(x)
}
\arguments{
\item{x}{A vector, can be character, numeric, or logical.}
}
\value{
A vector the same length as \code{x}, with all NAs or ""s replace by the last value for the vector.
Note that and missing values at the beginning of the vector will not be replaced.
}
\description{
Fill in missing values in a vector, using the last recorded value.
}
\examples{
numvec <- c(NA, 1:5, NA, NA, NA, 10:12, NA)
fill(numvec)

charvec <- c("", letters[1:5], "", "", "", letters[10:12], "")
fill(charvec)
}
\references{
Thomas A. Louis (1981). Confidence intervals for a binomial parameter after observing no successes.
The American Statistician, 35(3), 154.
\url{http://amstat.tandfonline.com/doi/abs/10.1080/00031305.1981.10479337?journalCode=utas20#.U3t2EvldX64}.
}

