% Generated by roxygen2 (4.0.1): do not edit by hand
\name{first}
\alias{first}
\title{Identify the First Elements of Series of Repeated Values}
\usage{
first(x)
}
\arguments{
\item{x}{A vector whose values will be explored for series of repeated values, can be character, numeric, or factor.}
}
\value{
An integer vector the same length as \code{x}, with a 1 for every element that is different than the one before it,
and a 0 for every element that is the same as the one before it.
}
\description{
Identify the first elements of series of repeated values.
}
\examples{
first(c(1, 2, 1, 2, 2, 1, 1, 3))
}

