% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatdf.R
\name{formatdf}
\alias{formatdf}
\title{Format a Data Frame}
\usage{
formatdf(df, numercol = NULL, round2 = 0, ndec = round2, comma = ",",
  characol = NULL, align = "left", keepnum = TRUE)
}
\arguments{
\item{df}{Data frame to be formatted.}

\item{numercol}{Numeric vector, index of numeric columns to format, default, NULL,
uses all numeric and integer columns.}

\item{round2}{Integer vector, indicating what place numbers should be rounded to,
default 0.}

\item{ndec}{Integer vector, indicating how many places to the right of
the decimal point should be displayed, default \code{round2}.}

\item{comma}{Character vector, indicating thousands character separator, default ",".}

\item{characol}{Numeric vector, index of character columns to format, default, NULL,
uses all character and factor columns.}

\item{align}{Character vector, indicating justification of character columns,
default "left". Other options include "right", "centre", and "none".}

\item{keepnum}{Logical scalar, indicating if numeric columns should be kept as numeric,
default TRUE, or converted to character.}
}
\value{
A data frame, the same dimensions as \code{df} with selected columns
  rounded and/or formatted and converted to character.
}
\description{
Format selected columns of a data frame for pretty viewing or printing.
}
\details{
If a vector of length 1 is provided for \code{round2}, \code{ndec},
  \code{comma}, or \code{align}, the same rounding, decimals, commas, or
  alignment is applied to all specified columns.
}
\examples{
head(mtcars)
formatdf(head(mtcars))
}
\seealso{
\code{\link{format}}.
}
