% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getpkgs.R
\name{getpackages}
\alias{getpackages}
\title{Get Packages}
\usage{
getpackages(want)
}
\arguments{
\item{want}{A character vector of package names.}
}
\description{
Installs (if necessary) and attaches the specified packages.
}

