% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpkgs.R
\name{getpkgs}
\alias{getpkgs}
\title{Get Packages}
\usage{
getpkgs(want, ...)
}
\arguments{
\item{want}{A character vector of package names.}

\item{...}{Additional arguments to \code{\link{library}}.}
}
\description{
Installs (if necessary) and attaches the specified packages.
}

