% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heading.r
\name{heading}
\alias{heading}
\title{Add a Heading to an RTF Document}
\usage{
heading(words, htype = 1, rtf = doc)
}
\arguments{
\item{words}{Character scalar text of heading to add to document.}

\item{htype}{Integer scalar heading type, 1 = bold and font size 12, 2 = bold and font size 10, 3 = italics and font size 10, default 1.}

\item{rtf}{An rtf object, default \code{doc}.}
}
\description{
Add a text heading to an rtf (rich text format) document.
}
\details{
The specified heading is written to the rtf file.
}
\examples{
\dontrun{
doc <- startrtf()
heading("Heading 1")
para("First paragraph.")
tab <- head(cars)
tabl("First few rows of cars data.", row.names=FALSE)
heading("Heading 2", 2)
para("Second paragraph.")
fig <- function() {
	plot(cars)
	lo <- loess(cars$dist ~ cars$speed)
	lines(lo$x, lo$fitted)
	}
figu("Speed vs. distance from the cars data.")
endrtf()
}
}
\seealso{
\code{\link[jvamisc]{startrtf}}, \code{\link[jvamisc]{para}}, \code{\link[jvamisc]{tabl}}, \code{\link[jvamisc]{figu}},
\code{\link[jvamisc]{figbig}}, \code{\link[jvamisc]{endrtf}}, \code{\link[rtf]{RTF}}.
}

