% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrange.R
\name{inrange}
\alias{inrange}
\title{In Range}
\usage{
inrange(x, r, open = TRUE)
}
\arguments{
\item{x}{A numeric vector of values to be tested.}

\item{r}{A numeric vector of length 2 specifying the range.}

\item{open}{A logical scalar indicating if the range interval is open
 (excludes endpoints, TRUE, default) or
closed (includes endpoints, FALSE).}
}
\value{
A logical vector, the same length as \code{x} indicating if values fall
  within the specified range.
}
\description{
Test to see if value falls within specified range.
}
\examples{
inrange(4:6, c(4, 5.5))
inrange(4:6, c(4, 5.5), FALSE)
}
