% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jvamisc-package.R
\docType{package}
\name{jvamisc}
\alias{jvamisc}
\alias{jvamisc-package}
\title{A Collection of Utility Functions}
\description{
\pkg{jvamisc} is a collection of miscellaneous utility functions.
Some of the more commonly used functions include
\code{\link{plotdf}} to plot each variable of a data frame,
\code{\link{shadepoly}} to add a shaded polygon of confidence intervals
to a plot,
\code{\link{AICc}} to compare models using Akaike's Information Criterion,
and \code{\link{startrtf}}, \code{\link{heading}}, \code{\link{para}},
\code{\link{tabl}}, \code{\link{figu}}, and \code{\link{endrtf}}
to write text, tables, and figures to a Word document.
}

