% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvatree.R
\name{jvatree}
\alias{jvatree}
\title{Cross Validated Tree Model}
\usage{
jvatree(...)
}
\arguments{
\item{...}{Arguments provided to \code{\link{rpart}} including formula
(the only required argument).}
}
\value{
An object of class rpart. See \code{\link{rpart.object}}.
}
\description{
Fit a classification or regression tree model to data, and
automatically prune the tree based on the cross-validation error.
}
\details{
First a tree model is fit.  Then the model is pruned such that it has
  the fewest number of splits that yield a mean cross-validation error
  less than the minimum cross-validation error plus one standard error.
}
\examples{
fit <- jvatree(Ozone ~ ., data=airquality)
par(xpd=NA) # otherwise on some devices the text is clipped
plot(fit)
text(fit, use.n=TRUE)
}

