% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{Identify the Last Elements of Series of Repeated Values}
\usage{
last(x)
}
\arguments{
\item{x}{A vector whose values will be explored for series of repeated values,
can be character, numeric, or factor.}
}
\value{
An integer vector the same length as \code{x}, with a 1 for every element
  that is different than the one after it, and a 0 for every element
  that is the same as the one after it.
}
\description{
Identify the last elements of series of repeated values.
}
\examples{
last(c(1, 2, 1, 2, 2, 1, 1, 3))
}
