% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong2.R
\name{latlong2}
\alias{latlong2}
\title{Convert Latitude and Longitude to US State or County}
\usage{
latlong2(pointsDF, to = c("state", "county")[1])
}
\arguments{
\item{pointsDF}{A data frame in which the first two columns contain the longitude and
latitude in degrees.}

\item{to}{A character scalar identifying the region to be identified, either "state",
the default, or "county".}
}
\value{
If \code{to="state"}, a character vector of state names, all lower case,
  the same length as the number of rows in \code{pointsDF}.
  If \code{to="county"}, a character matrix of state and county names, all
  lower case, with two columns and the same number of rows
  as \code{pointsDF}.
}
\description{
Convert latitude and longitude coordinates to US states or US states
and counties.
}
\examples{
testPoints <- data.frame(x=c(-90, -120), y=c(44, 44), z=c("a", "b"))
latlong2(testPoints, to="state")
latlong2(testPoints, to="county")
}
\references{
Based on a method posted by Josh O'Brien on 6 Jan 2012 on stackoverflow
  \href{http://stackoverflow.com/a/8751965/2140956}{[link]}.
}

