% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lls}
\alias{lls}
\title{List All Members of an Environment.}
\usage{
lls(pos = 1, pat = "")
}
\arguments{
\item{pos}{A scalar of the environment, search path index, or package of interest, default 1.}

\item{pat}{Regular expression pattern specifying which members to return, default "" for all members.}
}
\value{
A data frame with the name, class, dimension, and size of each member of the environment.
}
\description{
Generate a descriptive list of all members of an environment.
}
\examples{
\dontrun{
lls()
}
}
\references{
Based on a method posted by Bendix Carstensen on 10 Jan 2007 on R-help
\href{https://stat.ethz.ch/pipermail/r-help/2007-January/123403.html}{[link]}.
}

