% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong2utm.R
\name{longlat2utm}
\alias{longlat2utm}
\title{Convert Lat/Long to UTM}
\usage{
longlat2utm(lon, lat, zone)
}
\arguments{
\item{lon}{Numeric vector of longitudes.}

\item{lat}{Numeric vector of latitudes, same length as \code{lon}.}

\item{zone}{Numeric scalar of the UTM zone (see details).}
}
\value{
A data frame with two columns (easting and northing) and as many rows as
  the length of \code{lon} containing the converted UTM coordinates in
  meters.
}
\description{
Convert longitude and latitude coordinates to UTM easting and northing
coordinates.
}
\details{
For locations in the North American Great Lakes, zone 15 contains the
  Minnesota and Wisconsin part of Lake Superior; zone 16 has has rest of
  Lake Superior, all of Lake Michigan, and a small portion of western
  Lake Huron; zone 17 has rest of Lake Huron, all of Lake Erie, and
  the western half of Lake Ontario; and zone 18 has rest of Lake Ontario.
}
\examples{
longitude <- c(-82.27, -83.42, -81.01)
latitude <- c(44.76, 45.41, 45.45)
longlat2utm(longitude, latitude, 17)

}
\references{
Based on a function posted by Stanislav on 13 May 2015 on stackoverflow
  \href{http://stackoverflow.com/questions/18639967/converting-latitude-and-longitude-points-to-utm}{[link]}.
}

