% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-data.R
\name{mapSMR}
\alias{mapSMR}
\title{St. Marys River Shoreline}
\format{One row for each point, consecutive rows with non-missing values for
  each line.}
\source{
NOAA National Geophysical Data Center, (the now defunct) Coastline
  Extractor,
  \href{http://www.ngdc.noaa.gov/mgg/coast/}{ngdc.noaa.gov/mgg/coast}.
}
\description{
Data frame with longitude and latitude coordinates
  of the St. Marys River shoreline (the connecting channel between
  Lakes Superior and Huron in the Great Lakes basin).
}
\author{
Rich Signell, USGS.
}

