% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matrixtrim}
\alias{matrixtrim}
\title{Trim Matrix}
\usage{
matrixtrim(m, prop = c(1, 1), rowsfirst = TRUE)
}
\arguments{
\item{m}{A logical matrix to be trimmed, should be all TRUE, FALSE (or 0, 1).}

\item{prop}{A numeric vector of length two, the minimum proportion of TRUE values in each row and column of the trimmed matrix,
default c(1, 1).}

\item{rowsfirst}{A logical scalar, indicating if rows (TRUE, the default) or columns (FALSE) should be trimmed from the matrix first.}
}
\value{
A list of length three:
\itemize{
  \item \code{trim} = a logical matrix, the resulting trimmed matrix
  \item \code{dim} = a numeric vector of length 2, the dimensions of \code{trim}
  \item \code{n} = a numeric scaler, the total number of TRUE values in \code{trim}
}
}
\description{
Trim the rows and columns of a logical matrix until every row and columns has a specified level of TRUE values.
}
\examples{
m <- matrix(c(1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0), nrow=5)
matrixtrim(m)
matrixtrim(m, rowsfirst=FALSE)
matrixtrim(m, prop=c(0.7, 0.7))
}

