% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modecontin.R
\name{modecontin}
\alias{modecontin}
\title{Mode of Continuous Variable}
\usage{
modecontin(x, plot = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{plot}{A logical scalar indicating whether to plot the results, default FALSE.}

\item{...}{Additional arguments to the \code{\link{density}} function.}
}
\value{
The estimated mode of \code{x}.
}
\description{
Estimates the mode of a continuous variable.
}
\examples{
x <- rnorm(100)
modecontin(x, TRUE)
}
\references{
Based on a method posted by Peter Dalgaard on 29 Aug 2008 on
  \href{https://stat.ethz.ch/pipermail/r-help/2008-August/172319.html}{r-help}.
}

