% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvAntilog.R
\name{mvAntilog}
\alias{mvAntilog}
\title{Unbiased Prediction of Anti-Log Transformed Mean and Variance}
\usage{
mvAntilog(aovfit, xdata, out = c("predmean", "predvar", "predcv")[1])
}
\arguments{
\item{aovfit}{An object of class c("aov", "lm").}

\item{xdata}{A data frame with predictor variables corresponding to those in
\code{model} for which predictions should be made.}

\item{out}{A character scalar indicating what prediction value to return, either
"predmean" (default), "predvar", or "predcv".}
}
\value{
A numeric vector of predicted values on the original scale of the
  response.
}
\description{
Provide unbiased estimates of the mean and the variance
on the original scale from an analysis of
variance model with a natural log transformed response.
}
\examples{
fit <- aov(log(yield) ~ block + N * P + K, npk)
mvAntilog(fit, npk)
}
