% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mychi}
\alias{mychi}
\title{Chi-squared test}
\usage{
mychi(x, rpct = 0)
}
\arguments{
\item{x}{a numeric matrix}

\item{rpct}{a numeric scalar indicating the rounding used for printed output, default 0}
}
\value{
a list with class "htest" containing the components described in \link{\code{chisq.test}}
}
\description{
Performs chi-squared contingency table tests with informative output.
}
\examples{
## From Agresti(2007) p.39
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(gender = c("M", "F"), party = c("Democrat", "Independent", "Republican"))
mychi(M)
}
\seealso{
\link{\code{chisq.test}}
}

