% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Convert Integers to Words}
\usage{
numbers2words(x, billion = c("US", "UK")[1], and = if (billion == "US") ""
  else "and")
}
\arguments{
\item{x}{An integer vector to be converted.}

\item{billion}{A character scalar indicating if "US" (default) or "UK" meaning of
billion should be used.}

\item{and}{A character scalar for use as a conjunction, set to "" when
billion=="US" (default) and "and" otherwise.}
}
\value{
A character vector, the same length as \code{x} giving the English word(s)
  for the integer(s) \code{x}.
}
\description{
Convert integers into English words.
}
\examples{
numbers2words(c(4560000000, -123, 1000, 12))
}
\references{
John Fox.  2005.
  How Do You Spell That Number?
  Rnews Volume 5(1):51-54.
  \href{http://cran.r-project.org/doc/Rnews/Rnews_2005-1.pdf}{[link]}.
}

