% Generated by roxygen2 (4.0.1): do not edit by hand
\name{para}
\alias{para}
\title{Add a Paragraph to an RTF Document}
\usage{
para(..., rtf = doc, bold = FALSE, italic = FALSE)
}
\arguments{
\item{...}{One or more character scalars (separated by commas) of text to add to document as a single paragraph.}

\item{rtf}{An rtf object, default \code{doc}.}

\item{bold}{Logical scalar indicating if paragraph should use bold font, default FALSE.}

\item{italic}{Logical scalar indicating if paragraph should use italic font, default FALSE.}
}
\description{
Add a paragraph to an rtf (rich text format) document.
}
\details{
The specified heading is written to the rtf file.
}
\examples{
\dontrun{
doc <- startrtf()
heading("Heading 1")
para("First paragraph.")
tab <- head(cars)
tabl("First few rows of cars data.", row.names=FALSE)
heading("Heading 2", 2)
para("Second paragraph.")
fig <- function() {
	plot(cars)
	lo <- loess(cars$dist ~ cars$speed)
	lines(lo$x, lo$fitted)
	}
figu("Speed vs. distance from the cars data.")
endrtf()
}
}
\seealso{
\code{\link[jvamisc]{startrtf}}, \code{\link[jvamisc]{heading}}, \code{\link[jvamisc]{tabl}}, \code{\link[jvamisc]{figu}},
\code{\link[jvamisc]{figbig}}, \code{\link[jvamisc]{endrtf}}, \code{\link[rtf]{RTF}}.
}

