% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkgin.R
\name{pkgin}
\alias{pkgin}
\title{Install a Package}
\usage{
pkgin(package, wd = "C:/JVA/R/Working Directory",
  ld = "C:/Users/jvadams/Documents/R/win-library/3.1", pd = "C:/JVA/GitHub")
}
\arguments{
\item{package}{A character scalar, package name.}

\item{wd}{A character scalar, R working directory,
default "C:/JVA/R/Working Directory".}

\item{ld}{A character scalar, R library directory,
default "C:/Users/jvadams/Documents/R/win-library/3.1".}

\item{pd}{A character scalar, R package directory, default "C:/JVA/GitHub".}
}
\description{
Install a package from local files, load the library, and
save installed package to zip archive.
}
\seealso{
\code{\link{pkgup}}
}

