% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkgman.R
\name{pkgman}
\alias{pkgman}
\title{Copy a Package Manual}
\usage{
pkgman(package, file = paste0("C:/Users/jvadams/", package, ".Rcheck/",
  package, "-manual.pdf"), dir = "C:/JVA/GitHub/")
}
\arguments{
\item{package}{A character scalar, package name.}

\item{file}{A character scalar, path of file to be copied, default
paste0("C:/Users/jvadams/", package, ".Rcheck/", package, "-manual.pdf").}

\item{dir}{A character scalar, directory where package folder is located, default
"C:/JVA/GitHub/".}
}
\description{
Copy a package reference manual created by R CMD Check to the corresponding
GitHub folder.
}
\seealso{
\code{\link{pkgin}}
}

