% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkgup.R
\name{pkgup}
\alias{pkgup}
\title{Document a Package}
\usage{
pkgup(package, dir = "C:/JVA/GitHub/")
}
\arguments{
\item{package}{A character scalar, package name.}

\item{dir}{A character scalar, directory where package folder is located,
default "C:/JVA/GitHub/".}
}
\description{
Document a package from local files.
}
\seealso{
\code{\link{pkgin}}
}

