% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotblank.R
\name{plotblank}
\alias{plotblank}
\title{Create a Blank Plot}
\usage{
plotblank(x = 0:1, y = 0:1, xlab = "", ylab = "", las = 1, ...)
}
\arguments{
\item{x}{A numeric vector, the x coordinates of points in the plot.}

\item{y}{A numeric vector, the y coordinates of points in the plot.}

\item{xlab}{A character scalar, title for the x axis, default "".}

\item{ylab}{A character scalar, title for the y axis, default "".}

\item{las}{A numeric scalar, style of axis labels, 0=always parallel to the axis,
1=always horizontal (default), 2=always perpendicular to the axis,
3=always vertical.}

\item{...}{Additional arguments to the \code{\link{plot}} function.}
}
\description{
Create a blank plot (no symbols) on which to add other plotting features.
}
\examples{

plotblank(xlim=c(1, 100))
}
\seealso{
\code{\link{plot}}, \code{\link{title}}, \code{\link{par}}
}
