% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotcor}
\alias{plotcor}
\title{Plot a Correlation Matrix}
\usage{
plotcor(r, addtext = TRUE, incdiag = FALSE, rorder = TRUE, ...)
}
\arguments{
\item{r}{A symmetric correlation matrix with values ranging from -1 to 1.}

\item{addtext}{A logical scalar indicating whether the value of each correlation should be written over the ellipses, default TRUE.}

\item{incdiag}{A logical scalar indicating whether to plot circles for the diagonal values of the matrix, default FALSE.}

\item{rorder}{A logical scalar indicating whether the columns and rows of the matrix should be reordered using seriation, default TRUE.}

\item{...}{Additional parameters to \code{\link{par}}.}
}
\value{
A vector of integers (the same length as each dimensions of \code{r}) representing the linear order suggested by seriation.
}
\description{
Plot a correlation matrix using white ellipses (representing the measured correlation) overlaid on colored circles.
}
\details{
Each ellipse is sized so that the proportion of the colored circle visible beyond the ellipse is equal to the squared correlation.
Each ellipse is oriented northeast for positive correlation and northwest for negative correlation.
The color of each circle ranges from cyan (for correlation = -1) to magenta (for correlation = 1) through white
(for correlation = 0).  Similarly, the transparency of each correlation value (if addtext=TRUE) ranges from 1 (for correlation = 0) to
0 (for absolute correlation = 1).
}
\examples{
plotcor(cor(swiss))
plotcor(cor(longley))
}
\seealso{
\code{\link[ellipse]{plotcorr}}, on which the idea for the function was based, and \code{\link{seriation}}.
}

