% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predAntilog.R
\name{predAntilog}
\alias{predAntilog}
\title{Unbiased Prediction of Log Transformed Response on Original Scale}
\usage{
predAntilog(aovfit, xdata, logbase = exp(1), k = 0)
}
\arguments{
\item{aovfit}{An object of class c("aov", "lm").}

\item{xdata}{A data frame with predictor variables corresponding to those in
\code{model} for which predictions should be made.}

\item{logbase}{A numeric scalar, the base of the log transformation used in
the transformed response of \code{model}, default exp(1).}

\item{k}{A numeric scalar, the constant added to the response prior to
transformation, default 0.}
}
\value{
A numeric vector of predicted values on the original scale of the
  response.
}
\description{
Provide unbiased estimates on the original scale from an analysis of
variance model with a log transformed response.
}
\examples{
fit <- aov(log(yield) ~ block + N * P + K, npk)
predAntilog(fit, npk)
}
