% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prettytable.R
\name{prettytable}
\alias{prettytable}
\title{Prettify the Numeric Columns of a Table}
\usage{
prettytable(m, sigdig = 3)
}
\arguments{
\item{m}{Two dimensional data frame, matrix, or table to be prettified.}

\item{sigdig}{Integer scalar, number of significant digits to be used, default 3.}
}
\value{
A data frame, matrix, or table, the same dimensions as \code{m} with
  the numeric columns rounded to \code{sigdig} significant digits.
}
\description{
Prettify the numeric columns of a table for printing.
}
\examples{
head(mtcars)
prettytable(head(mtcars), 2)
}
\seealso{
\code{\link{signif}}.
}

