% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/q.r
\name{q}
\alias{q}
\title{Quit}
\usage{
q(save = "yes")
}
\arguments{
\item{save}{A character string indicating whether the workspace should be saved, "yes" (the default), "no", "ask" or "default".}
}
\description{
Terminate the current R session.
}
\seealso{
\code{\link{quit}}.
}

