% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recode}
\alias{recode}
\title{Recode Values}
\usage{
recode(x, old, new, must.match = TRUE)
}
\arguments{
\item{x}{A vector whose values will be recoded, can be character, numeric, or factor.}

\item{old}{A vector of the unique values currently in the vector.}

\item{new}{A vector of values which should replace the current ones.}

\item{must.match}{A logical scalar indicating whether only those elements of the original vector with values in \code{old} should be returned (TRUE),
or all values should be returned (FALSE) though some may be unchanged, default TRUE.}
}
\value{
A vector the same length as \code{x} (unless \code{must.match=TRUE}), with \code{old} values replaced by \code{new} values.
}
\description{
Recode the values of a vector.
}
\examples{
recode(c(1,1,1,2,3,4,1,10,3), 1:3, 1001:1003)
recode(c(1,1,1,2,3,4,1,10,3), 1:3, 1001:1003, must.match=F)
}

