% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/segreg.R
\name{segreg}
\alias{segreg}
\title{Segmented Regression}
\usage{
segreg(x, y, k = NULL)
}
\arguments{
\item{x}{Numeric vector, the independent variable which will be "broken".}

\item{y}{Numeric vector, the dependent variable which will be used to define
the break point.}

\item{k}{Numeric scalar, the location of the break point, if known.
If NULL, the default, the optimal break point will be chosen from
among all unique values of x except for the minimum and the maximum.}
}
\value{
List with three elements, fit: the resulting lm object, k, and pred:
  data frame with the unique values of x (and k) with predicted values.
}
\description{
Fit a simple linear segmented regression (also known as changepoint,
hockey stick, or broken line regression).
}
\details{
The break point cuts the x vector into two groups, x <= k and x > k.
}
\examples{
indvar <- sample(1:50, 100, replace=TRUE)
depvar <- ifelse(indvar<32, indvar + 3, indvar/4 + 27) + rnorm(100, sd=2)
sr <- segreg(indvar, depvar)
plot(indvar, depvar)
lines(sr$pred)
abline(v=sr$k, lty=2)
}

