% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shadepoly}
\alias{shadepoly}
\title{Add Shaded Polygon to Plot}
\usage{
shadepoly(x, ymd, ylo, yhi, subsel = NULL, kol = "#000000", opq = c(20,
  50), addline = TRUE)
}
\arguments{
\item{x}{A numeric vector, the metric to plot in the x direction (e.g., time).}

\item{ymd}{A numeric vector, the metric to plot in the y direction (e.g., a mean or median).}

\item{ylo}{A numeric vector, the lower interval or range in the y direction (e.g., a lower confidence interval or quartile).}

\item{yhi}{A numeric vector, the upper interval or range in the y direction (e.g., an upper confidence interval or quartile).}

\item{subsel}{A logical vector, indicating subset of the data to plot.}

\item{kol}{A character scalar, the hex color to use for plotting both the shaded polygon and (if requested) the line, default "#000000" (black).}

\item{opq}{A numeric vector of length 2, opacity for the polygon and the line, default c(20, 50).}

\item{addline}{A logical scalar, indicating if the x vs ymd line should be added to the plot (on top of the shaded polygon), default TRUE.}
}
\description{
Add a spline-smoothed, shaded polygon to a plot, typically to show an interval or range of values (in the y-direction) for a time series of x values.
}
\details{
Missing values are removed prior to plotting, such that there will be no breaks in the shaded polygon nor the line (if requested).
The lower and upper intervals of the polygon are spline-smoothed prior to plotting, as is the line (if requested).
}
\examples{
x <- 1:10
y <- sample(10)
noise <- abs(rnorm(10))
plot(x, y, ylim=range(y-noise, y+noise), type="n")
shadepoly(x, y, y-noise, y+noise)
}
\seealso{
\code{\link{polygon}}, \code{\link{lines}}.
}

