% Generated by roxygen2 (4.0.2): do not edit by hand
\name{showmarks}
\alias{showmarks}
\title{Show Marks}
\usage{
showmarks()
}
\description{
Show marks used in graphing, including line types, plotting symbols, default colors, color blind friendly colors (`blindcolz`), and fonts.
}
\examples{
showmarks()
}

