% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/startrtf.R
\name{startrtf}
\alias{startrtf}
\title{Create an RTF Document}
\usage{
startrtf(file = NULL, dir = getwd(), width = 8.5, height = 11,
  omi = c(1, 1, 1, 1), quiet = FALSE)
}
\arguments{
\item{file}{Character scalar name of document, default "RGeneratedDocument" with
\code{\link{Sys.Date}} suffix.}

\item{dir}{Character scalar name of directory where document should be stored,
default \code{\link{getwd}()}.}

\item{width}{Numeric scalar width of document page in inches, default 8.5.}

\item{height}{Numeric scalar height of document page in inches, default 11.}

\item{omi}{Numeric vector, length 4, width of document page margins in inches
(bottom, left, top, right), default c(1, 1, 1, 1).}

\item{quiet}{Logical scalar indicating if name of new rtf document should be printed to
command line, default FALSE.}
}
\value{
An rtf file is created in the specified directory.
  An object of class rtf is created.  This object is referred to in
  other function to write to the file.
  In addition, two numeric vectors of length 1, \code{tabcount} and
  \code{figcount}, are written to the working directory to keep track
  of the number of tables and figures written to the rtf document, and
  label the captions accordingly.
}
\description{
Create an rtf (rich text format) document.
}
\details{
The rtf file may be written to until the \code{\link{endrtf}()} function
  is run.
  If you assign your rtf file to an object called \code{doc},
  you can use the default settings in other \pkg{\link{jvamisc}}
  rtf functions.
}
\examples{
\dontrun{
doc <- startrtf()
heading("Heading 1")
para("First paragraph.")
tab <- head(cars)
tabl("First few rows of cars data.", row.names=FALSE)
heading("Heading 2", 2)
para("Second paragraph.")
fig <- function() {
	plot(cars)
	lo <- loess(cars$dist ~ cars$speed)
	lines(lo$x, lo$fitted)
}
figu("Speed vs. distance from the cars data.")
endrtf()
}
}
\seealso{
\code{\link[jvamisc]{heading}}, \code{\link[jvamisc]{para}},
  \code{\link[jvamisc]{tabl}}, \code{\link[jvamisc]{figu}},
  \code{\link[jvamisc]{figbig}}, \code{\link[jvamisc]{endrtf}},
  \code{\link[rtf]{RTF}}.
}

