% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc-data.R
\name{states}
\alias{states}
\title{United States State Boundaries Map}
\format{A list (of class map) with elements x, y, range, and names.
  The x and y vectors have longitude (x) and latitude (y) coordinates of
  successive state polygons, separated by NAs.}
\source{
A copy of the \code{state} data base from the \code{maps} package.
}
\description{
Geographical data base for the state boundaries of the US,
  generated from US Department of the Census data.
}

