% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stringin.r
\name{stringin}
\alias{stringin}
\title{Find a String in a Character Vector}
\usage{
stringin(pattern, x, ignore.case = TRUE, value = TRUE, fixed = TRUE, ...)
}
\arguments{
\item{pattern}{Character scalar, string to be matched in \code{x}.}

\item{x}{Character vector where matches are sought.}

\item{ignore.case}{Logical scalar, indicating case sensitivity, default TRUE.}

\item{value}{Logical scalar, indicating whether to return the matching elements (the default, TRUE) or their indices (FALSE).}

\item{fixed}{Logical scalar, indicating whether string should be matched as is, default TRUE.}

\item{...}{Other arguments to \code{\link{grep}}.}
}
\value{
Character vector containing the matching elements of \code{x}.
}
\description{
Find a string in a character vector.
}
\examples{
txt <- c("The", "licenses", "for", "most", "software", "are", "designed",
	"to", "take", "away", "your", "freedom", "to", "share", "and", "change",
	"it.", "", "By", "contrast,", "the", "GNU", "General", "Public", "License",
	"is", "intended", "to", "guarantee", "your", "freedom", "to", "share", "and",
	"change", "free", "software", "--", "to", "make", "sure", "the", "software",
	"is", "free", "for", "all", "its", "users.")
stringin("b", txt)
stringin("ar", txt)
}
\seealso{
\code{\link{grep}}.
}

