% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subdex.R
\name{subdex}
\alias{subdex}
\title{Subset Index}
\usage{
subdex(x, subset)
}
\arguments{
\item{x}{Vector, matrix, or data frame to be subsetted.}

\item{subset}{Logical expression indicating elements or rows to keep;
missing values are taken as FALSE.}
}
\value{
Logical vector of indices,
  the same length as \code{x} (if \code{x} is a vector) or
  the rows of \code{x} (if \code{x} is a matrix or data frame).
}
\description{
Return the index of vectors, matrices or data frames which meet conditions.
}
\examples{
df <- data.frame(x=c(1, 2, NA, 4, 20), y=c(5, NA, 6, 30, 4))
subdex(df, x > 8 | y > 8)
df[subdex(df, x > 8 | y > 8), ]
# compare to subset
subset(df, x > 8 | y > 8)

m <- as.matrix(df)
subdex(m, x > 8 | y > 8)
m[subdex(m, x > 8 | y > 8), ]
# compare to subset
subset(m, m[, "x"] > 8 | m[, "y"] > 8)

z <- c(1, 2, NA, 4, 20)
subdex(z, z > 8)
z[subdex(z, z > 8)]
# compare to subset
subset(z, z > 8)
}
\seealso{
\code{\link{subset}}, from which the function was derived.
}

