% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tabl}
\alias{tabl}
\title{Add a Table to an RTF Document}
\usage{
tabl(..., TAB = tab, rtf = doc, fontt = 8, row.names = TRUE,
  tabc = jvamiscenv$tabcount, boldt = TRUE, newpage = "none", omi = c(1,
  1, 1, 1))
}
\arguments{
\item{...}{One or more character scalars (separated by commas) of text to use for the table caption.}

\item{TAB}{A matrix, data frame, or table to be added to the document as a table, default \code{tab}.}

\item{rtf}{An rtf object, default \code{doc}.}

\item{fontt}{Numeric scalar font size for table caption, default 8.}

\item{row.names}{Logical scalar whether to include the row.names of \code{TAB} in the table, default TRUE.}

\item{tabc}{Numeric scalar table number to use in caption, default \code{jvamiscenv$tabcount}.}

\item{boldt}{Logical scalar indicating if table number should use bold font, default TRUE.}

\item{newpage}{Character scalar indicating if the table should start on a new page in the document "port" for a new portrait page,
"land" for a new landscape page, and "none" for no new page (the default).}

\item{omi}{Numeric vector, length 4, width of document page margins in inches (bottom, left, top, right), default c(1, 1, 1, 1).}
}
\value{
A 1 is added to the numeric vector of length 1, \code{jvamiscenv$tabcount}, stored in the working directory to keep track
of the number of tables written to the rtf document, and label the captions accordingly.
}
\description{
Add a table to an rtf (rich text format) document.
}
\details{
The table and caption are written to the rtf file.
The size of a new page is assumed to be 8.5 by 11 inches.
}
\examples{
\dontrun{
doc <- startrtf()
heading("Heading 1")
para("First paragraph.")
tab <- head(cars)
tabl("First few rows of cars data.", row.names=FALSE)
heading("Heading 2", 2)
para("Second paragraph.")
fig <- function() {
	plot(cars)
	lo <- loess(cars$dist ~ cars$speed)
	lines(lo$x, lo$fitted)
	}
figu("Speed vs. distance from the cars data.")
endrtf()
}
}
\seealso{
\code{\link[jvamisc]{startrtf}}, \code{\link[jvamisc]{heading}}, \code{\link[jvamisc]{para}}, \code{\link[jvamisc]{figu}},
\code{\link[jvamisc]{figbig}}, \code{\link[jvamisc]{endrtf}}, \code{\link[rtf]{RTF}}.
}

