% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trimspace.R
\name{trimspace}
\alias{trimspace}
\title{Trim Whitespace}
\usage{
trimspace(charvec)
}
\arguments{
\item{charvec}{A character vector to be trimmed.}
}
\value{
A character vector, the same length as \code{charvec},
  without leading and trailing whitespace.
}
\description{
Trim leading and trailing whitespace from a character vector.
}
\examples{
trimspace("    Bob   and Alice")
trimspace(" Harriet and    June    ")
}

