% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweethead.R
\name{tweethead}
\alias{tweethead}
\title{Tweet Headlines}
\usage{
tweethead(posttweet = TRUE, username = NULL, website = NULL,
  credOrPath = "C:/JVA/R/Working Directory/.Renviron")
}
\arguments{
\item{posttweet}{A logical scalar indicating if tweets should be posted, default TRUE.}

\item{username}{A character scalar, giving the name of the twitter user. The default, NULL,
uses information stored in local .Renviron file (see Details).}

\item{website}{A character scalar, giving the name of the website, from which to pull
headlines. The default, NULL, uses information stored in local .Renviron
file (see Details).}

\item{credOrPath}{Either a character scalar giving the path of the environment where the
credentials are stored or a character vector of length four, giving the
credentials themselves: twitter_api_key, twitter_api_secret,
twitter_access_token, and twitter_access_token_secret.  The default is
"C:/JVA/R/Working Directory/.Renviron" (see Details).}
}
\value{
A list of length 3.  The first element is a data frame of old tweets
  with text, favoriteCount, retweetCount, and createdUTC (date/time) as
  columns.  The second and third elements are character strings of the
  current headlines (currentheads) and the latests items to tweet (totweet).
}
\description{
Tweet the latest headlines from the specified website.
\pkg{twitteR} and \pkg{RCurl} packages required.
}
\details{
This function is customized to work on a particular website.
  It's not for general use. To store information in local .Renviron file,
  use \code{writeLines(c("username=xxx", "website=xxx",
  "twitter_api_key=xxx", "twitter_api_secret=xxx",
  "twitter_access_token=xxx", "twitter_access_token_secret=xxx"),
  file.path(getwd(), ".Renviron"))}.
}
\examples{
\dontrun{
tweethead(FALSE)
tweethead()
}
}
\references{
Simon Munzert.  19 Jan 2015.
Programming a Twitter bot - and the rescue from procrastination.
\emph{http://www.r-datacollection.com/blog/Programming-a-Twitter-bot/}

Simon Munzert.  21 Dec 2014.
How to conduct a tombola with R.
\emph{www.r-datacollection.com/blog/How-to-conduct-a-tombola-with-R/}
}

