% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper2full.R
\name{upper2full}
\alias{upper2full}
\title{Convert Upper Triangular Matrix to Full Matrix}
\usage{
upper2full(up, diagval = NULL)
}
\arguments{
\item{up}{A matrix in which the upper trianglular part is of interest.}

\item{diagval}{A vector of values to be used for the diagonal of the returned
symmetric matrix.  May be the same length as each dimension of \code{up}
or may be of length 1.  If Null, the default, the diagonal values will
be taken from \code{up}.}
}
\value{
A symmetric matrix the same dimensions as \code{up}, with diagonal values
  assigned according to \code{diagval}.
}
\description{
Convert the upper triangular part of a matrix to a full symmetric matrix.
}
\examples{
m <- matrix(1:16, nrow=4, byrow=TRUE)
upper2full(m)
upper2full(m, 0)
upper2full(m, 1:4)
}

