% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedges.R
\name{wedges}
\alias{wedges}
\title{Add Wedge-Shaped Line Segments to a Plot}
\usage{
wedges(x0, y0, x1, y1, ybase = 0.03 * diff(par("usr")[3:4]), pborder = NA,
  pcol = "darkgray", ...)
}
\arguments{
\item{x0}{Numeric vector of x coordinates \strong{from} which to draw
the wide base of the wedges.}

\item{y0}{Numeric vector of y coordinates \strong{from} which to draw
the wide base of the wedges.  Must be the same length as \code{x0}.}

\item{x1}{Numeric vector of x coordinates \strong{to} which to draw
the narrow point of the wedges.  Must be the same length as \code{x0}.}

\item{y1}{Numeric vector of y coordinates \strong{to} which to draw
the narrow point of the wedges.  Must be the same length as \code{x0}.}

\item{ybase}{Numeric vector giving the width of the wedge bases in y-scale units,
default 3% of the y-axis range.  Must be of length 1 or the same
length as \code{x0}.}

\item{pborder}{Vector of colors to draw the borders of the wedges.
The default, NA, omits borders.
Must be of length 1 or the same length as \code{x0}.
See \code{\link{polygon}}.}

\item{pcol}{Vector of colors to fill the wedges, default "darkgray".
Must be of length 1 or the same length as \code{x0}.
See \code{\link{polygon}}.}

\item{...}{Additional arguments to \code{\link{polygon}}.}
}
\description{
Draw wedge-shaped line segments between pairs of points.
}
\examples{
df <- data.frame(x=runif(5, 0, 1), y=runif(5, 0, 10))

# draw a single wedge
plot(df)
with(df, wedges(x[1], y[1], x[2], y[2], ybase=1))

# draw several wedges from point to point
s <- seq(length(df$x)-1)
plot(df)
with(df, invisible(lapply(s, function(i)
 wedges(x[i], y[i], x[i+1], y[i+1], pborder=(5:2)[i], pcol=(2:5)[i])
)))

}
\seealso{
\code{\link{segments}}, \code{\link{arrows}}, \code{\link{polygon}}.
}
