% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCitations.R
\name{xCitations}
\alias{xCitations}
\title{Extract Citations from Text}
\usage{
xCitations(file = NULL, txt = NULL)
}
\arguments{
\item{file}{Character scalar indicating the name of the text file.  Default value of
NULL allows for text to be read in directly using the \code{txt} argument.}

\item{txt}{Character scalar containing the text containing citations.  Default value
of NULL expects text to be read in from a file using the \code{file}
argument.}
}
\value{
A character vector of unique citations in alphabetical order.
}
\description{
Extract citations from text by searching for parenthetical phrases with
capital words (names) and 4-digit numbers (years)
}
\details{
Citations in which the author is mentioned directly in the text and
  only the year is in parentheses are not captured with this function.
}
\examples{
x <- paste("Yarmouth (1977) said something.",
 "Evidence of x (Barber and Jones 1991),",
 "y (House et al. 1982),",
 "and z (Smith 1990; Folger and Penn 2000).")
xCitations(txt=x)

}
\references{
Based on code posted by Kay Cichini on 26 Mar 2012 on theBioBucket
  \href{http://thebiobucket.blogspot.com/2012/03/how-to-extract-citation-from-body-of.html}{[link]}.
}
