% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allcombs}
\alias{allcombs}
\title{All Combinations}
\usage{
allcombs(num, from = 0, to = num)
}
\arguments{
\item{num}{A scalar, the number of items.}

\item{from}{A scalar, the minimum number of items in each combination, default 0.}

\item{to}{A scalar, the maximum number of items in each combination, default \code{num}.}
}
\value{
A matrix with rows corresponding to each possible combination and columns corresponding to item number.
}
\description{
All possible combinations of the given number of items.
}
\examples{
allcombs(3)
}

