% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomCI.R
\name{binomCI}
\alias{binomCI}
\title{Binomial Confidence Interval}
\usage{
binomCI(x, y = NULL, na.rm = TRUE, alpha = 0.05, prob = TRUE)
}
\arguments{
\item{x}{A vector of 0s and 1s, or (if \code{y} is also given) a scalar,
the number of successes.}

\item{y}{A scalar, the number of failures, default \code{NULL}.}

\item{na.rm}{A logical, whether to remove NAs from the data, default \code{TRUE}.}

\item{alpha}{A scalar, the desired confidence level, default \code{0.05}.}

\item{prob}{A logical, whether to output results as probabilities,
default \code{TRUE}, or counts.}
}
\value{
A named vector with the \code{Mean}, lower and upper confidence limits
  (\code{L} and \code{U}), and the number of observations \code{N}.
}
\description{
Calculates the binomial confidence interval from a sample, using
the normal approximation.
Uses Louis (1981) if only failures or only successes were observed.
}
\examples{
binomCI(c(0, 0, 0, 0, 1, 1))
binomCI(2, 4, prob=FALSE)
}
\references{
Thomas A. Louis. 1981.
  Confidence intervals for a binomial parameter after observing no successes.
  The American Statistician 35(3):154.
  \emph{http://amstat.tandfonline.com/doi/abs/10.1080/00031305.1981.10479337?journalCode=utas20#.U3t2EvldX64}
}

