% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc-data.R
\name{blindcolz}
\alias{blindcolz}
\title{Color-blind Friendly Colors}
\format{A character vector, length 9.}
\source{
How to make figures and presentations that are friendly to color blind
  people,
  20 November 2002,
  \href{http://jfly.iam.u-tokyo.ac.jp/html/color_blind/}{[link]}.
}
\description{
A vector with nine color-blind friendly colors in hex code.
}
\author{
Masataka Okabe and Kei Ito.
}

