% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcr2}
\alias{calcr2}
\title{Coefficient of Determination}
\usage{
calcr2(fitted, observed, nparam)
}
\arguments{
\item{fitted}{A vector of fitted values.}

\item{observed}{A vector of observed values.}

\item{nparam}{A scalar, number of parameters in the fitted model.}
}
\value{
A vector of unadjusted and adjusted coefficients of determination.
}
\description{
Calculate the unadjusted and adjusted coefficient of determination (R^2).
}
\examples{
fit1 <- lm(mpg ~ cyl + disp, data=mtcars)
calcr2(fit1$fitted, mtcars$mpg, 3)

fit2 <- lm(mpg ~ cyl + disp + wt, data=mtcars)
calcr2(fit2$fitted, mtcars$mpg, 4)
}

