% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chi.R
\name{chi}
\alias{chi}
\title{Chi-Squared Test}
\usage{
chi(x, rpct = 0, print = TRUE, plot = TRUE)
}
\arguments{
\item{x}{A numeric matrix.}

\item{rpct}{A numeric scalar indicating the rounding used for printed output,
default 0.}

\item{print}{A logical scalar indicating if output should be printed, default TRUE.}

\item{plot}{A logical scalar indicating if plot should be generated, default TRUE.}
}
\value{
A list with class "htest" containing the components described in
  \code{\link{chisq.test}}.
}
\description{
Performs chi-squared contingency table tests with informative output.
}
\examples{
## From Agresti(2007) p.39
M <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(M) <- list(gender=c("M", "F"),
	party=c("Democrat", "Independent", "Republican"))
chi(M)
}
\seealso{
\code{\link{chisq.test}}.
}

