% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coordtri}
\alias{coordtri}
\title{Determine Third Coordinate of Triangle}
\usage{
coordtri(cleft, cright, lleft, lright)
}
\arguments{
\item{cleft}{A numeric vector of length 2, the x and y coordinates of the lower left point of the triangle.}

\item{cright}{A numeric vector of length 2, the x and y coordinates of the lower right point of the triangle.}

\item{lleft}{A numeric scalar, the length of the left side of the triangle, from \code{cleft} to the top.}

\item{lright}{A numeric scalar, the length of the right side of the triangle, from \code{cright} to the top.}
}
\value{
A numeric vector of length 2, the x and y coordinates of the top point of the triangle.
}
\description{
Determine the third coordinate of a triangle, given the other two coordinates and the distances to the third.
}
\examples{
# define coordinates of base of triangle
AB <- rbind(A=c(-2, 4), B=c(2, -4))
# determine coordinates of top of triangle, given base coordinates and side lengths
C <- coordtri(cleft=AB[1, ], cright=AB[2, ], lleft=12, lright=10)
# plot results
coordplot(rbind(C, AB))
}
\seealso{
\code{\link[jvamisc]{coordplot}}, \code{\link[jvamisc]{coordturn}}, \code{\link[jvamisc]{coordmove}}, \code{\link[jvamisc]{coordflip}}.
}

