% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dfclip.r
\name{dfclip}
\alias{dfclip}
\title{Read Data Frame from Clipboard}
\usage{
dfclip(...)
}
\arguments{
\item{...}{Additional parameters to \code{\link{read.table}}.}
}
\value{
Data frame.
}
\description{
Read in a data frame from text pasted to the clipboard.
}
\seealso{
\code{\link{read.table}}.
}

