% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfclip}
\alias{dfclip}
\title{Data frame from clipboard}
\usage{
dfclip(...)
}
\arguments{
\item{...}{parameters to \link{\code{read.table}}}
}
\value{
data frame
}
\description{
Read in a data frame from headers/rows pasted to the clipboard.
}

