% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fill}
\alias{fill}
\title{Fill in Missing Values}
\usage{
fill(x, resetwhen = rep(FALSE, length(x)))
}
\arguments{
\item{x}{A vector, can be character, numeric, or logical.}

\item{resetwhen}{A logical vector, the same length as \code{x}.}
}
\value{
A vector the same length as \code{x}, with all NAs or ""s replace by the last value for the vector.
Note that and missing values at the beginning of the vector will not be replaced.
}
\description{
Fill in missing values in a vector, using the last recorded value.
}
\details{
Similar to \code{\link[zoo]{na.locf}} in the \code{zoo} package, but works for "" in character vectors as well.
}
\examples{
numvec <- c(NA, 1:5, NA, NA, NA, 10:12, NA)
group <- rep(1:3, c(4, 4, 5))
fill(numvec)
fill(numvec, first(group))

charvec <- c("", letters[1:5], "", "", "", letters[10:12], "")
fill(charvec)
}

