% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvaFirst.R
\name{jvaFirst}
\alias{jvaFirst}
\title{Startup}
\usage{
jvaFirst(
  maxp = 500,
  ndec = 10,
  cont = "... ",
  mirror = "http://streaming.stat.iastate.edu/CRAN/",
  helpt = "html",
  fac = FALSE
)
}
\arguments{
\item{maxp}{Integer scalar, maximum number of lines printed, default 500.}

\item{ndec}{Integer scalar, maximum width of fixed notation before switching to
scientific notation, default 10.}

\item{cont}{Character scalar, prompt used for lines which continue past
first command line, default \code{"... "}.}

\item{mirror}{Character scalar, CRAN mirror, default
"http://streaming.stat.iastate.edu/CRAN/".}

\item{helpt}{Character scalar, type of help, default "html".}

\item{fac}{Logical scalar, use factors rather than character strings, default FALSE.}
}
\description{
One function with all the commands I typically want run at the start of
an R session.
}
\examples{
jvaFirst()
}
\seealso{
\code{\link{options}}, \code{\link{help}}.
}
