% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{jvamisc}
\alias{jvamisc}
\alias{jvamisc-package}
\title{A Collection of Miscellaneous Functions}
\description{
\pkg{jvamisc} is simply a collection of miscellaneous functions that I find useful.
Some of the more commonly used functions include
\code{\link{plotdf}} to plot each variable of a data frame,
\code{\link{shadepoly}} to add a shaded polygon of confidence intervals to a plot,
\code{\link{AICc}} to compare models using Akaike's Information Criterion,
and \code{\link{startrtf}}, \code{\link{heading}}, \code{\link{para}}, \code{\link{tabl}}, \code{\link{figu}}, and \code{\link{endrtf}} to
to write text, tables, and figures to a Word document.
}

