% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc-data.R
\name{jvamiscenv}
\alias{jvamiscenv}
\title{jvamisc package local environment}
\format{An environment.}
\source{
Post from Hadley Wickham to r-help on 2 Dec 2014
	\href{https://stat.ethz.ch/pipermail/r-help/2014-December/423847.html}{[link]}.
}
\description{
An environment local to the jvamisc package, used to hold objects
  outside of the individual package functions
	(e.g., the table and figure counts for the rtf functions).
}

