% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvanames.R
\name{jvanames}
\alias{jvanames}
\title{Format Names}
\usage{
jvanames(charvec)
}
\arguments{
\item{charvec}{A character vector to be trimmed.}
}
\value{
A character vector, the same length as \code{charvec}, but with
  all lower case, unique, and without underscores.
}
\description{
Format names to be all lower case, unique, and without underscores.
}
\examples{
jvanames(c("Conc_mgL", "Conc_mgL", "Temp"))
}
\seealso{
\code{\link{make.names}}
}
