% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{map5}
\alias{map5}
\title{Great Lakes Basin Shoreline}
\format{One row for each point, consecutive rows with non-missing values for each line.}
\source{
NOAA National Geophysical Data Center, (the now defunct) Coastline
	Extractor, \href{http://www.ngdc.noaa.gov/mgg/coast/}{ngdc.noaa.gov/mgg/coast}.
}
\usage{
map5
}
\description{
A single data frame with longitude and latitude coordinates
of the shorelines in the Great Lakes basin.
}
\author{
Rich Signell, USGS.
}

