% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-data.R
\name{mapL}
\alias{mapL}
\title{Great Lakes Shorelines}
\format{One row for each point, consecutive rows with non-missing values for
  each line.}
\source{
NOAA National Geophysical Data Center, (the now defunct) Coastline
  Extractor,
  \href{http://www.ngdc.noaa.gov/mgg/coast/}{ngdc.noaa.gov/mgg/coast}.
}
\description{
List of five data frames, one for each Great Lake
  (Superior, Michigan, Huron, Erie, Ontario)
  with longitude and latitude coordinates of the shorelines.
}
\author{
Rich Signell, USGS.
}
