% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mrc.r
\name{mrc}
\alias{mrc}
\title{Estimates from Mark-Recapture Studies}
\usage{
mrc(m, r, c, alpha = 0.05)
}
\arguments{
\item{m}{A numeric scalar, the number of individuals marked and released.}

\item{r}{A numeric scalar, the number of marked individuals recaptured.}

\item{c}{A numeric scalar, the number of individuals checked for marks.}

\item{alpha}{A scalar, the desired confidence level, default \code{0.05}.}
}
\value{
A named vector with the estimated recovery rate \code{U} with lower and upper 100*(1-\code{alpha})\% confidence limits from \code{\link{binomCI}},
and the number of observations \code{N} (the bias-corrected Petersen estimator from Chapman 1954 Table 3)
with lower and upper 100*(1-\code{alpha})\% confidence limits from the normal approximation of the Poisson for \code{r} (Ricker 1975).
}
\description{
Estimates the recovery rate and abundance of a closed population from a mark-recapture study with a single capture and marking event and a single recapture event.
}
\examples{
mrc(100, 10, 1000)
}
\references{
D. G. Chapman.  1954.
	The estimation of biological populations.
	Annals of Mathematical Statistics 25:1-15.
\href{http://projecteuclid.org/euclid.aoms/1177728844}{[link]}.

W. E. Ricker.  1975.
	Computation and interpretation of biological statistics of fish populations.
	Fisheries Research Board of Canada Bulletin. 191.
}
\seealso{
\code{\link{binomCI}}
}

