% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgin.R
\name{pkgin}
\alias{pkgin}
\title{Install a Package}
\usage{
pkgin(package, ld = "C:/Users/jvadams/Documents/R/win-library", lv = "max",
  pd = "C:/JVA/GitHub")
}
\arguments{
\item{package}{A character scalar, package name.}

\item{ld}{A character scalar, R library directory,
default "C:/Users/jvadams/Documents/R/win-library/3.2".}

\item{lv}{A character scalar, R library directory version number,
default "max", which uses the latest version in \code{ld}.}

\item{pd}{A character scalar, R package directory, default "C:/JVA/GitHub".}
}
\description{
Install a package from local files and load the library.
}
\seealso{
\code{\link{pkgup}}
}

