% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgin.R
\name{pkgin}
\alias{pkgin}
\title{Install a Package}
\usage{
pkgin(package, ld = getwd(), pd = "C:/JVA/GitHub")
}
\arguments{
\item{package}{A character scalar, package name.}

\item{ld}{A character scalar, current local directory, default \code{getwd()}.}

\item{pd}{A character scalar, R package directory, default "C:/JVA/GitHub".}
}
\description{
Install a package from local files and load the library.
\pkg{devtools} package required.
}
\seealso{
\code{\link{pkgup}}
}
