% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgup}
\alias{pkgup}
\title{Document a Package}
\usage{
pkgup(package, dir = "C:/JVA/GitHub/")
}
\arguments{
\item{package}{A character scalar, package name.}

\item{dir}{A character scalar, directory where package folder is located, default "C:/JVA/GitHub/".}
}
\description{
Document a package from local files.
}
\seealso{
\code{\link{pkgin}}
}

