% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotcor}
\alias{plotcor}
\title{Plot a Matrix of Correlations}
\usage{
plotcor(r, addtext = TRUE, atcex = NULL, incdiag = FALSE, rorder = TRUE,
  ...)
}
\arguments{
\item{r}{A matrix of correlations (need not be symmetrical) with values ranging from -1 to 1.}

\item{addtext}{A logical scalar indicating whether the value of each correlation should be written over the ellipses, default TRUE.}

\item{atcex}{A numeric scalar giving the magnification of the added ellipse text relative to that set in \code{\link{par}},
default NULL results in atcex = 8/maximum dimension of \code{r}.}

\item{incdiag}{A logical scalar indicating whether to plot circles for the diagonal values of the matrix (if symmetric), default FALSE.}

\item{rorder}{A logical scalar indicating whether the columns and rows of the matrix should be reordered using seriation, default TRUE.}

\item{...}{Additional parameters to \code{\link{par}}.}
}
\value{
A list of with two vector of integers (the same length as each dimension of \code{r}) representing the linear order suggested by seriation.
}
\description{
Plot a matrix of correlations using white ellipses (representing the measured correlation) overlaid on colored circles.
}
\details{
Each ellipse is sized so that the proportion of the colored circle visible beyond the ellipse is equal to the squared correlation.
Each ellipse is oriented northeast for positive correlation and northwest for negative correlation.
The color of each circle ranges from cyan (for correlation = -1) to magenta (for correlation = 1) through white
(for correlation = 0).  Similarly, the transparency of each correlation value (if addtext=TRUE) ranges from 1 (for correlation = 0) to
0 (for absolute correlation = 1).
}
\examples{
# example using a symmetric matrix
sr <- cor(swiss)
sord <- plotcor(sr)
sr[sord[[1]], sord[[2]]]
# example using an asymmetric matrix
lr <- cor(longley)[1:3, 4:7]
lord <- plotcor(lr)
lr[lord[[1]], lord[[2]]]
}
\seealso{
\code{\link[ellipse]{plotcorr}}, on which the idea for the function was based, \code{\link{seriation}},
	and \code{\link[plotrix]{draw.ellipse}}.
}

