% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotdf.R
\name{plotdf}
\alias{plotdf}
\title{Plot Data Frame}
\usage{
plotdf(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Other parameters to \code{par(...)}.}
}
\value{
A data frame with summary statistics for each variable in the data frame,
  number entered, number missing, number unique, minimum, mean, maximum, and
  the ratio of the maximum over the non-zero minimum.
}
\description{
Plots each variable of a data frame.
}
\examples{
plotdf(mtcars)
}
\seealso{
\code{\link{par}}.
}

