% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotdf}
\alias{plotdf}
\title{Plot data frame}
\usage{
plotdf(df, one = TRUE, ...)
}
\arguments{
\item{df}{a data frame}

\item{one}{a logical, whether to display all plots in a single \code{windows(record=TRUE)} device, default TRUE}

\item{...}{other parameters to \code{par(...)}}
}
\value{
a data frame with summary statistics for each variable in the data frame, number entered, number missing, number unique, minimum, mean, maximum, and
the maximum over the non-zero minimum
}
\description{
Plots each variable of a data frame.
}
\examples{
plotdf(mtcars)
}
\seealso{
\code{\link{windows}}, \code{\link{par}}
}

