% Generated by roxygen2 (4.0.0): do not edit by hand
\name{q}
\alias{q}
\title{Quit}
\usage{
q()
}
\arguments{
\item{save}{a character string indicating whether the workspace should be saved, "yes" (the default), "no", "ask" or "default"}
}
\value{
the R session is terminated (nothing is returned)
}
\description{
Quit R
}
\seealso{
\code\link{{quit}}
}

