% Generated by roxygen2 (4.0.1): do not edit by hand
\name{q}
\alias{q}
\title{Quit}
\usage{
q()
}
\arguments{
\item{save}{A character string indicating whether the workspace should be saved, "yes" (the default), "no", "ask" or "default".}
}
\description{
Terminate the current R session.
}
\seealso{
\code{\link{quit}}.
}

