% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/showmarks.r
\name{showmarks}
\alias{showmarks}
\title{Show Marks}
\usage{
showmarks()
}
\description{
Show marks used in graphing, including line types, plotting symbols, default colors, color blind friendly colors (`blindcolz`), and fonts.
}
\examples{
showmarks()
}

