% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stratCochran.R
\name{stratCochran}
\alias{stratCochran}
\title{Stratified Random Sampling Estimates}
\usage{
stratCochran(yhi, hi, Wh = NULL, N = NULL)
}
\arguments{
\item{yhi}{A numeric vector of values obtained from the ith unit (in stratum h).}

\item{hi}{A numeric vector of stratum ids, the same length as yhi.}

\item{Wh}{A numeric vector of stratum weights, with length equal to the number
of unique stratum ids (nh), ordered by stratum id.  If NULL (the default),
weights are assumed to be equal for all strata (1/nh).}

\item{N}{Total of the quantities used for the stratum weights across all strata,
used to expand estimated means to estimated totals.  If NULL (the
default), no totals are estimated.}
}
\value{
A named numeric vector of length 4, with the estimated mean and total,
  and their associated standard errors.
}
\description{
Calculates estimated means and variances from a stratified random
sampling design.
}
\examples{
catch <- c(2, 4, 2, 4, 8, 9, 8, 9, 8, 9)
stratum <- c(1, 1, 1, 1, 2, 2, 2, 2, 2, 2)
area <- c(3, 7)
stratCochran(yhi=catch, hi=stratum, Wh=area/sum(area), N=sum(area))
stratCochran(yhi=catch, hi=stratum)
}
\references{
Cochran, William G.  1977.
  Sampling Techniques.
  John Wiley & Sons, New York.
}

