% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweethead.R
\name{tweethead}
\alias{tweethead}
\title{Tweet Headlines}
\usage{
tweethead(posttweet = TRUE, username = NULL, website = NULL)
}
\arguments{
\item{posttweet}{A logical scalar indicating if tweets should be posted, default TRUE.}

\item{username}{A character scalar, giving the name of the twitter user. The default, NULL,
uses information stored in local .Renviron file.}

\item{website}{A character scalar, giving the name of the website, from which to pull
headlines. The default, NULL, uses information stored in local .Renviron
file.}
}
\value{
A list of length 3.  The first element is a data frame of old tweets
  with text, favoriteCount, retweetCount, and createdUTC (date/time) as
  columns.  The second and third elements are character strings of the
  current headlines (currentheads) and the latests items to tweet (totweet).
}
\description{
Tweet the latest headlines from the specified website.
\pkg{rtweet} and \pkg{RCurl} packages required.
}
\details{
This function is customized to work on a particular website.
  It's not for general use.
}
\examples{
\dontrun{
tweethead(FALSE)
tweethead()
}
}
\references{
Simon Munzert.  19 Jan 2015.
Programming a Twitter bot - and the rescue from procrastination.
\emph{http://www.r-datacollection.com/blog/Programming-a-Twitter-bot/}

Simon Munzert.  21 Dec 2014.
How to conduct a tombola with R.
\emph{www.r-datacollection.com/blog/How-to-conduct-a-tombola-with-R/}
}
